Tidy-up & Migration files.
Split out sub/inner classes from ImagesController.cs into own class files.

Migrations files for remapping state and status for screens and reviews.
