+++ Inspicio/Models/ReviewViewModels/DataFromToggle.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace Inspicio.Controllers
+{
+    public partial class ImagesController : Controller
+    {
+        public class DataFromToggle
+        {
+            public int ReviewId { get; set; }
+            public bool Open { get; set; }
+
+        }
+    }
+}
