+++ Inspicio/Models/ReviewViewModels/SelectableUser.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Inspicio.Models;
+
+namespace Inspicio.Controllers
+{
+    public partial class ImagesController : Controller
+    {
+        public class SelectableUser : ApplicationUser
+        {
+            public bool IsSelected { get; set; }
+        }
+    }
+}
